% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkmode.R
\name{darkmode}
\alias{darkmode}
\title{Enable DarkMode with default button}
\usage{
darkmode(
  session = getDefaultReactiveDomain(),
  bottom = "32px",
  right = "32px",
  left = "unset",
  time = "0.5s",
  mixColor = "#fff",
  backgroundColor = "#fff",
  buttonColorDark = "#100f2c",
  buttonColorLight = "#fff",
  saveInCookies = FALSE,
  label = "<U+0001F313>",
  autoMatchOsTheme = TRUE
)
}
\arguments{
\item{session}{Session}

\item{label}{button label for toggling dark mode}
}
\value{

}
\description{
It lets user to toggle darkmode with default button
}
\examples{
library(shiny)
library(shinyDarkmode)
vars <- setdiff(names(iris), "Species")
ui <- fluidPage(

  titlePanel("Sample App"),

  # Setup
  use_darkmode(),

  sidebarPanel(
    selectInput('xcol', 'X Variable', vars),
    selectInput('ycol', 'Y Variable', vars, selected = vars[[2]]),
    numericInput('clusters', 'Cluster count', 3, min = 1, max = 9)
  ),
  mainPanel(
    plotOutput('plot1')
  ),

  column(width = 12,
         DT::dataTableOutput("mydt")
  )
)

# Server
server <- function(input, output, session) {

  darkmode(label = "⏳")

  # Combine the selected variables into a new data frame
  selectedData <- reactive({
    iris[, c(input$xcol, input$ycol)]
  })

  clusters <- reactive({
    kmeans(selectedData(), input$clusters)
  })

  output$plot1 <- renderPlot({
    palette(c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3",
              "#FF7F00", "#FFFF33", "#A65628", "#F781BF", "#999999"))

    par(mar = c(5.1, 4.1, 0, 1))
    plot(selectedData(),
         col = clusters()$cluster,
         pch = 20, cex = 3)
    points(clusters()$centers, pch = 4, cex = 4, lwd = 4)
  })

  output$mydt <- DT::renderDataTable(iris,
                                     options = list(pageLength = 10, autoWidth = TRUE),
                                     rownames= FALSE)

}

# Run App
shinyApp(ui = ui, server = server)
}
\author{
Deepanshu Bhalla
}
