% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkmode.R
\name{darkmode_toggle}
\alias{darkmode_toggle}
\title{Toggle darkmode with custom checkbox}
\usage{
darkmode_toggle(
  session = getDefaultReactiveDomain(),
  inputid = "togglemode",
  saveInCookies = FALSE,
  autoMatchOsTheme = TRUE
)
}
\arguments{
\item{session}{CurrentSession}

\item{inputid}{Input ID of checkbox (or switch in shinyWidgets)}

\item{saveInCookies}{whether save in cookies}

\item{autoMatchOsTheme}{BMatch rowser Theme}
}
\value{

}
\description{
Toggle darkmode with custom checkbox
}
\examples{
library(shiny)
library(shinyWidgets)
library(shinyDarkmode)
ui <- fluidPage(

  titlePanel("Sample App"),

  # Setup
  use_darkmode(),

  # Inputs
  fluidRow(
    tags$div(style = "margin-top: -25px; float: right; margin-right: -150px;",
             prettySwitch("togglemode", "Dark Mode", value = FALSE, fill = TRUE, status = "info")
    )
  ), DT::dataTableOutput("mydt")

)


# Server
server <- function(input, output, session) {

  darkmode_toggle(inputid = 'togglemode')

  output$mydt <- DT::renderDataTable(mtcars,
                                     options = list(pageLength = 10, autoWidth = TRUE),
                                     rownames= FALSE)

}

# Run App
shinyApp(ui = ui, server = server)
}
